ai/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
supabase==2.11.0
pandas==2.2.3
numpy==2.2.2
scipy==1.15.1
scikit-learn==1.6.1
joblib==1.4.2
pytz==2025.1
python-dateutil==2.9.0.post0
